fix friends request button

+++ backend/controller/friendController.js
@@ -1,8 +1,11 @@
 const User = require("../models/userModel");
 const Friend = require("../models/friendModel");
+const AppError = require("../utilities/appError");
 
 exports.getFriends = async (req, res, next) => {
   try {
+    if (!req.user.id) return next(new AppError("Please provide user id", 400));
+
     const results = await User.findById(req.user.id)
       .populate({ path: "friends", options: { sort: { status: 1 } } })
       .select(["-__v", "-introduction"])
@@ -24,6 +27,9 @@ exports.request = async (req, res, next) => {
     const userA = req.user.id;
     const userB = id;
 
+    if (!userA || !userB)
+      return next(new AppError("Please provide user id", 400));
+
     const docA = await Friend.findOneAndUpdate(
       { requester: userA, recipient: userB },
       { status: 1 },
@@ -35,6 +41,9 @@ exports.request = async (req, res, next) => {
       { status: 2 },
       { upsert: true, new: true }
     );
+
+    if (!docA || !docB)
+      return next(new AppError("No document found with that ID", 404));
 
     const updateUserA = await User.findOneAndUpdate(
       { _id: userA },
@@ -65,6 +74,9 @@ exports.accept = async (req, res, next) => {
     const userA = req.user.id;
     const userB = id;
 
+    if (!userA || !userB)
+      return next(new AppError("Please provide user id", 400));
+
     await Friend.findOneAndUpdate(
       { requester: userA, recipient: userB },
       { status: 3 }
@@ -90,6 +102,9 @@ exports.cancel = async (req, res, next) => {
     const userA = req.user.id;
     const userB = id;
 
+    if (!userA || !userB)
+      return next(new AppError("Please provide user id", 400));
+
     const docA = await Friend.findOneAndRemove({
       requester: userA,
       recipient: userB,
@@ -99,6 +114,9 @@ exports.cancel = async (req, res, next) => {
       requester: userB,
       recipient: userA,
     });
+
+    if (!docA || !docB)
+      return next(new AppError("No document found with that ID", 404));
 
     await User.findOneAndUpdate(
       { _id: userA },
